% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab03.R
\name{dijkstra}
\alias{dijkstra}
\title{Implement the dijkstra algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph, }{A data frame with three variables, v1, v2 and w that contains the edges of the graph (fromv1 to v2) with the weight of the edge (w)}

\item{init_node, }{A number.}
}
\description{
The function returns the shortest path to every other node in \code{graph} from \code{init_node}.
}
