% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lab03.R
\name{euclidean}
\alias{euclidean}
\title{Implement the Euclidean algorithm}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\description{
The greatest common divisor of \code{x} and \code{y}.
}
\references{
More information of euclidean algorithm \href{http://en.wikipedia.org/wiki/Euclidean}{here}
}
